requirements.txt
click==8.1.7
Flask==2.2.5
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.0.0
SQLAlchemy==2.0.20
typing_extensions==4.12.2
waitress==2.1.2
Werkzeug==3.1.3
